+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## [Unreleased]
 * Fixed
   * Configuration variable for copyright notice fixed.
+* Added
+  * Service added to optimize handling of EXTCONF values.
 * Changed
   * Password field changed to password type.
   * Configuration variable for update URI moved to _ext_localconf.php_.